Linux persistent code to linux/arch.c
